time: avoid printf-format warnings
Force 'long int' type for '%ld' specifiers (e.g. OpenBSD's time_t is not
'long int').

* src/time.c: (summarize) Type case time_t to lont int.

+++ src/time.c
@@ -474,14 +474,14 @@ summarize (fp, fmt, command, resp)
 	    case 'E':		/* Elapsed real (wall clock) time.  */
 	      if (resp->elapsed.tv_sec >= 3600)	/* One hour -> h:m:s.  */
 		fprintf (fp, "%ld:%02ld:%02ld",
-			 resp->elapsed.tv_sec / 3600,
-			 (resp->elapsed.tv_sec % 3600) / 60,
-			 resp->elapsed.tv_sec % 60);
+			 (long int)(resp->elapsed.tv_sec / 3600),
+			 (long int)((resp->elapsed.tv_sec % 3600) / 60),
+			 (long int)(resp->elapsed.tv_sec % 60));
 	      else
 		fprintf (fp, "%ld:%02ld.%02ld",	/* -> m:s.  */
-			 resp->elapsed.tv_sec / 60,
-			 resp->elapsed.tv_sec % 60,
-			 resp->elapsed.tv_usec / 10000);
+			 (long int)(resp->elapsed.tv_sec / 60),
+			 (long int)(resp->elapsed.tv_sec % 60),
+			 (long int)(resp->elapsed.tv_usec / 10000));
 	      break;
 	    case 'F':		/* Major page faults.  */
 	      fprintf (fp, "%ld", resp->ru.ru_majflt);
@@ -493,9 +493,10 @@ summarize (fp, fmt, command, resp)
         case 'K': /* Average mem usage == data+stack+text.  */
           fprintf (fp, "%lu",
                    MSEC_TO_TICKS (v) == 0 ? 0 :
-                   get_rusage_idrss_kb (&resp->ru) / MSEC_TO_TICKS (v) +
-                   get_rusage_isrss_kb (&resp->ru) / MSEC_TO_TICKS (v) +
-                   get_rusage_ixrss_kb (&resp->ru) / MSEC_TO_TICKS (v));
+		   (long unsigned int)
+                   (get_rusage_idrss_kb (&resp->ru) / MSEC_TO_TICKS (v) +
+                    get_rusage_isrss_kb (&resp->ru) / MSEC_TO_TICKS (v) +
+                    get_rusage_ixrss_kb (&resp->ru) / MSEC_TO_TICKS (v)));
           break;
         case 'M': /* Maximum resident set size.  */
           fprintf (fp, "%" PRIuMAX, get_rusage_maxrss_kb (&resp->ru));
@@ -516,13 +517,13 @@ summarize (fp, fmt, command, resp)
 	      break;
 	    case 'S':		/* System time.  */
 	      fprintf (fp, "%ld.%02ld",
-		       resp->ru.ru_stime.tv_sec,
-		       resp->ru.ru_stime.TV_MSEC / 10);
+		       (long int)resp->ru.ru_stime.tv_sec,
+		       (long int)(resp->ru.ru_stime.TV_MSEC / 10));
 	      break;
 	    case 'U':		/* User time.  */
 	      fprintf (fp, "%ld.%02ld",
-		       resp->ru.ru_utime.tv_sec,
-		       resp->ru.ru_utime.TV_MSEC / 10);
+		       (long int)(resp->ru.ru_utime.tv_sec),
+		       (long int)(resp->ru.ru_utime.TV_MSEC / 10));
 	      break;
 	    case 'W':		/* Times swapped out.  */
 	      fprintf (fp, "%ld", resp->ru.ru_nswap);
@@ -542,8 +543,8 @@ summarize (fp, fmt, command, resp)
 	      break;
 	    case 'e':		/* Elapsed real time in seconds.  */
 	      fprintf (fp, "%ld.%02ld",
-		       resp->elapsed.tv_sec,
-		       resp->elapsed.tv_usec / 10000);
+		       (long int)resp->elapsed.tv_sec,
+		       (long int)(resp->elapsed.tv_usec / 10000));
 	      break;
 	    case 'k':		/* Signals delivered.  */
 	      fprintf (fp, "%ld", resp->ru.ru_nsignals);
